refactor: cli error
