+++ packages/repack/src/commands/common/config/getConfigFilePath.ts
@@ -6,7 +6,7 @@ import {
   DEFAULT_RSPACK_CONFIG_LOCATIONS,
   DEFAULT_WEBPACK_CONFIG_LOCATIONS,
 } from '../../consts.js';
-import { CLIError } from '../error.js';
+import { CLIError } from '../cliError.js';
 
 function discoverConfigFilePath(root: string, candidates: string[]) {
   for (const candidate of candidates) {
