+++ packages/repack/src/commands/rspack/Compiler.ts
@@ -11,7 +11,7 @@ import type {
 import memfs from 'memfs';
 import type { Reporter } from '../../logging/types.js';
 import type { HMRMessage } from '../../types.js';
-import { CLIError } from '../common/error.js';
+import { CLIError } from '../common/cliError.js';
 import { adaptFilenameToPlatform, runAdbReverse } from '../common/index.js';
 import { DEV_SERVER_ASSET_TYPES } from '../consts.js';
 import type { CompilerAsset } from './types.js';
