+++ packages/repack/src/commands/webpack/start.ts
@@ -10,8 +10,8 @@ import {
   makeLogEntryFromFastifyLog,
 } from '../../logging/index.js';
 import type { HMRMessage } from '../../types.js';
+import { CLIError } from '../common/cliError.js';
 import { makeCompilerConfig } from '../common/config/makeCompilerConfig.js';
-import { CLIError } from '../common/error.js';
 import {
   getMimeType,
   parseFileUrl,
